Waypoint 2 - all systems working
Need to test
1. Form2dmap stars working again - SystemClass has a func to return all
star positions
2. VisitedSystemsClass - moved some worker functions to there out of
other classes
3. 3D Formmap now prefers to find a system on ID not on name, if it has
an ID.  This means duplicates will work better
4. long is used for id's in DB - id, id_edsm and id_eddb.
5. edsm_id introduced to system class
6. System table has an edsm_id field.  Two new index to system table
id_edsm and id_eddb.
7. On DB upgrade to 17, force a EDSM and EDDB sync (by clearing their
time values, see UpgradeDB17)
8. SystemClass supports id_edsm. Delete/update more flexible about id
use.  New functions to support other items
9. EDSM parser now used id_edsm to sync.  All items in db not having an
id_edsm are removed.
Hidden system parser asks for the id_edsm to know which to remove - does
not rely on name
So, EDSM sync is now not named based. Duplicate names can now be
supported
10. EDDB parser now looks for the EDDB Id to sync with, if that fails
looks for name at the approx correct location
if no match, no data. else data is updated
11. Savedrout control now working again. autocomplete just used the
global class instead of being clever
12. Travelhistory - moved some funcs to visitedsystemclass
13. Trilat works - suggested references works, using a call back from
systemdata to populate its sectors. limits
stars per sector to 1000 to keep memory use down
Trilat shows xyz on suggested systems (nicety).  Autocomplete just uses
global autocomplete.
The old expression was too slow to work with this db system

+++ EDDiscovery/DB/IVisitedSystems.cs
@@ -4,7 +4,7 @@ namespace EDDiscovery2.DB
 {
     public interface IVisitedSystems
     {
-        int id { get; set; }
+        long id { get; set; }
         string Name { get; set; }
         DateTime Time { get; set; }
         int Commander { get; set; }
